Reorganize front page markup and styles
Beginning with the page before login, I changed the most awkward
arrangements of the page markup, and simplified the CSS that controlled
those parts.

New CSS is being added to main.css.scss

This commit is large because I ran a formatter over two of the CSS files
so that I could read them while I was working on them.